+++ Infrastructure/SecurityManagers/AspNetIdentity/IdentityService.cs
@@ -8,6 +8,7 @@ using Application.Services.Repositories;
 using AutoMapper;
 using Domain.Constants;
 using Domain.Entities;
+using Domain.Interfaces;
 using Infrastructure.DataAccessManagers.EFCores.Contexts;
 using Infrastructure.SecurityManagers.RoleClaims;
 using Infrastructure.SecurityManagers.Tokens;
@@ -119,6 +120,11 @@ namespace Infrastructure.SecurityManagers.AspNetIdentity
             var tokens = _queryContext.Token.Where(t => t.UserId == userId).ToList();
             _queryContext.Token.RemoveRange(tokens);
             _queryContext.SaveChanges();
+
+            var deleteResult = await _photoService.DeletePhotoAsync(user.ProfilePictureName);
+            if (deleteResult == null)
+                throw new ApplicationException("Xóa ảnh không thành công");
+
             var result = await _userManager.DeleteAsync(user);
 
             if (!result.Succeeded)
